# Fix unsafe modification of Trampoline resources