--- a/monix-execution/js/src/main/scala/monix/execution/schedulers/TrampolineExecutionContext.scala
+++ b/monix-execution/js/src/main/scala/monix/execution/schedulers/TrampolineExecutionContext.scala
@@ -18,7 +18,7 @@
 package monix.execution.schedulers
 
 import monix.execution.internal.Trampoline
-import monix.execution.internal.Trampoline.{ForkingEC, ImmediateEC, TrampolineEC}
+import monix.execution.internal.Trampoline.{ForkingTrampolineEC, ImmediateTrampolineEC, TrampolineEC}
 
 import scala.concurrent.{ExecutionContext, ExecutionContextExecutor}
 
@@ -70,7 +70,7 @@ object TrampolineExecutionContext {
     *        is needed
     */
   def apply(underlying: ExecutionContext): TrampolineExecutionContext =
-    new TrampolineExecutionContext(new ForkingEC(underlying))
+    new TrampolineExecutionContext(new ForkingTrampolineEC(underlying))
 
   /** [[TrampolineExecutionContext]] instance that executes everything
     * immediately, on the current thread.
@@ -86,5 +86,5 @@ object TrampolineExecutionContext {
     *    have no way to override it)
     */
   val immediate: TrampolineExecutionContext =
-    new TrampolineExecutionContext(ImmediateEC)
+    new TrampolineExecutionContext(ImmediateTrampolineEC)
 }