Forked executions of Trampoline tasks operated on the same data structures as the thread-local Trampoline. Consequently, multiple threads were able to concurrently access same non-thread-safe resources.

Fixed by operating solely on copied data in the forked Trampoline instead of referencing the previous-thread-local Trampoline.