@mikkolaj 2025-07-14
Shallow execution (`3.3.0`):
```
Result "monix.benchmarks.TrampolinedRunnableBenchmark.shallow":
  14532.889 ±(99.9%) 189.720 ops/s [Average]
  (min, avg, max) = (13952.961, 14532.889, 14781.169), stdev = 218.481
  CI (99.9%): [14343.169, 14722.608] (assumes normal distribution)

Benchmark                             (size)   Mode  Cnt      Score     Error  Units
TrampolinedRunnableBenchmark.shallow    3000  thrpt   20  14532.889 ± 189.720  ops/s
```

Shallow execution (latest, after fix):
```
Result "monix.benchmarks.TrampolinedRunnableBenchmark.shallow":
  14429.338 ±(99.9%) 266.398 ops/s [Average]
  (min, avg, max) = (13624.400, 14429.338, 14744.336), stdev = 306.784
  CI (99.9%): [14162.940, 14695.736] (assumes normal distribution)

Benchmark                             (size)   Mode  Cnt      Score     Error  Units
TrampolinedRunnableBenchmark.shallow    3000  thrpt   20  14429.338 ± 266.398  ops/s
```
Deep execution (`3.3.0`):
```
Result "monix.benchmarks.TrampolinedRunnableBenchmark.deep":
  11790.037 ±(99.9%) 622.659 ops/s [Average]
  (min, avg, max) = (8801.810, 11790.037, 12124.772), stdev = 717.055
  CI (99.9%): [11167.378, 12412.696] (assumes normal distribution)

Benchmark                          (size)   Mode  Cnt      Score     Error  Units
TrampolinedRunnableBenchmark.deep    3000  thrpt   20  11790.037 ± 622.659  ops/s
```

Deep execution (latest, after fix):
```
Result "monix.benchmarks.TrampolinedRunnableBenchmark.deep":
  11447.485 ±(99.9%) 510.602 ops/s [Average]
  (min, avg, max) = (10099.265, 11447.485, 12115.118), stdev = 588.010
  CI (99.9%): [10936.882, 11958.087] (assumes normal distribution)

Benchmark                          (size)   Mode  Cnt      Score     Error  Units
TrampolinedRunnableBenchmark.deep    3000  thrpt   20  11447.485 ± 510.602  ops/s
```