@mikkolaj 2025-07-28
Updated benchmarks after the latest change:

Benchmark before (`3.4.2-avs.4`):
```
Result "monix.benchmarks.TrampolinedRunnableBenchmark.deep":
  24930.680 ±(99.9%) 579.269 ops/s [Average]
  (min, avg, max) = (23615.111, 24930.680, 25963.704), stdev = 667.088
  CI (99.9%): [24351.411, 25509.950] (assumes normal distribution)

Result "monix.benchmarks.TrampolinedRunnableBenchmark.shallow":
  23275.552 ±(99.9%) 1714.118 ops/s [Average]
  (min, avg, max) = (18440.319, 23275.552, 25220.304), stdev = 1973.981
  CI (99.9%): [21561.434, 24989.671] (assumes normal distribution)
  
Result "monix.benchmarks.TrampolineExecutionContextBenchmark.immediateBlocking":
  21539.378 ±(99.9%) 198.276 ops/s [Average]
  (min, avg, max) = (21154.088, 21539.378, 22043.127), stdev = 228.334
  CI (99.9%): [21341.102, 21737.653] (assumes normal distribution)

Benchmark                                              (size)   Mode  Cnt      Score      Error  Units
TrampolinedRunnableBenchmark.deep                        3000  thrpt   20  24930.680 ±  579.269  ops/s
TrampolinedRunnableBenchmark.shallow                     3000  thrpt   20  23275.552 ± 1714.118  ops/s
TrampolineExecutionContextBenchmark.immediateBlocking    3000  thrpt   20  21539.378 ±  198.276  ops/s
```

Benchmark after:
```
Result "monix.benchmarks.TrampolinedRunnableBenchmark.deep":
  25423.862 ±(99.9%) 467.468 ops/s [Average]
  (min, avg, max) = (24016.334, 25423.862, 26067.530), stdev = 538.337
  CI (99.9%): [24956.394, 25891.330] (assumes normal distribution)

Result "monix.benchmarks.TrampolinedRunnableBenchmark.shallow":
  24601.949 ±(99.9%) 201.177 ops/s [Average]
  (min, avg, max) = (24091.664, 24601.949, 25074.417), stdev = 231.676
  CI (99.9%): [24400.772, 24803.126] (assumes normal distribution)

Result "monix.benchmarks.TrampolineExecutionContextBenchmark.immediateBlocking":
  23377.368 ±(99.9%) 516.149 ops/s [Average]
  (min, avg, max) = (21859.586, 23377.368, 24199.765), stdev = 594.398
  CI (99.9%): [22861.218, 23893.517] (assumes normal distribution)

Benchmark                                              (size)   Mode  Cnt      Score     Error  Units
TrampolinedRunnableBenchmark.deep                        3000  thrpt   20  25423.862 ± 467.468  ops/s
TrampolinedRunnableBenchmark.shallow                     3000  thrpt   20  24601.949 ± 201.177  ops/s
TrampolineExecutionContextBenchmark.immediateBlocking    3000  thrpt   20  23377.368 ± 516.149  ops/s
```